change: ニックネームを入力する場所を変更する
