+++ package.json
@@ -29,6 +29,7 @@
   "devDependencies": {
     "@testing-library/jest-dom": "^6.1.4",
     "@testing-library/react": "^14.0.0",
+    "eslint-plugin-tailwindcss": "^3.13.0",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0"
   }
